LayoutTests/http/tests/security/cross-frame-access-history-get-expected.txt

CONSOLE MESSAGE: line 55: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: TypeError: undefined is not an object (evaluating 'window.targetWindow.history.length')
CONSOLE MESSAGE: line 55: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: TypeError: undefined is not an object (evaluating 'window.targetWindow.history.back')
CONSOLE MESSAGE: line 55: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: TypeError: undefined is not an object (evaluating 'window.targetWindow.history.forward')
CONSOLE MESSAGE: line 55: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: TypeError: undefined is not an object (evaluating 'window.targetWindow.history.go')
CONSOLE MESSAGE: line 55: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: TypeError: undefined is not an object (evaluating 'window.targetWindow.history.toString')
CONSOLE MESSAGE: line 55: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 55: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 55: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: TypeError: undefined is not an object (evaluating 'window.targetWindow.history.existingCustomProperty')
CONSOLE MESSAGE: line 55: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: TypeError: undefined is not an object (evaluating 'window.targetWindow.history.__proto__')
CONSOLE MESSAGE: line 94: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 97: TypeError: undefined is not an object (evaluating 'window.targetWindow.history.existingCustomFunction')
CONSOLE MESSAGE: line 94: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 97: TypeError: undefined is not an object (evaluating 'window.targetWindow.history.prototypeCustomFunction')


----- tests for getting window.history and its properties -----

PASS: canGet('targetWindow.history.length') should be 'false' and is.
PASS: canGet('targetWindow.history.back') should be 'false' and is.
PASS: canGet('targetWindow.history.forward') should be 'false' and is.
PASS: canGet('targetWindow.history.go') should be 'false' and is.
PASS: canGet('targetWindow.history.toString') should be 'false' and is.
PASS: canGet('targetWindow.__proto__') should be 'false' and is.
PASS: canGet('targetWindow.constructor') should be 'false' and is.
PASS: canGet('targetWindow.history.existingCustomProperty') should be 'false' and is.
PASS: canGet('targetWindow.history.__proto__.prototypeCustomProperty') should be 'false' and is.
PASS: canCall('targetWindow.history.existingCustomFunction') should be 'false' and is.
PASS: canCall('targetWindow.history.prototypeCustomFunction') should be 'false' and is.
